# docs/swagger.yaml
basePath: /api/v1
definitions:
  response.Data:
    properties:
      code:
        $ref: '#/definitions/response.ResCode'
      data: {}
      msg: {}
    type: object
  response.ResCode:
    enum:
    - 20000
    format: int64
    type: integer
    x-enum-varnames:
    - CodeSuccess
...
  schema.CategoryBasicInfo:
    properties:
      id:
        description: Category ID | 版块ID
        example: 1
        type: integer
      name:
        description: Category name | 版块名称
        example: 综合讨论
        type: string
      slug:
        description: Category slug | 版块标识
        example: general
        type: string
    type: object
...
  schema.UserCreateRequest:
    properties:
      avatar:
        description: Avatar URL | 头像URL
        example: https://example.com/avatar.jpg
        type: string
      email:
        description: Email | 邮箱
        example: test@example.com
        type: string
      password:
        description: Password | 密码
        example: password123
        minLength: 8
        type: string
      readme:
        description: README content | README内容
        example: |-
          ## 关于我
          这是我的README内容
        type: string
      role:
        description: User role | 用户身份
        enum:
        - User
        - Moderator
        - Admin
        - SuperAdmin
        example: User
        type: string
      signature:
        description: Signature | 个性签名
        example: 这是我的个性签名
        type: string
      username:
        description: Username | 用户名
        example: testuser
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - password
    - role
    - username
    type: object
...
  schema.UserDetailResponse:
    properties:
      avatar:
        description: Avatar URL | 头像URL
        example: https://example.com/avatar.jpg
        type: string
      comment_count:
        description: Comment count | 评论数
        example: 200
        type: integer
      created_at:
        description: Creation time | 创建时间
        example: "2024-01-01 00:00:00"
        type: string
      currency:
        description: Currency | 货币
        example: 500
        type: integer
      email:
        description: Email | 邮箱
        example: test@example.com
        type: string
      email_verified:
        description: Whether email verified | 邮箱是否已验证
        example: true
        type: boolean
      id:
        description: User ID | 用户ID
        example: 1
        type: integer
      managed_categories:
        description: Managed categories list (only shown for moderators) | 管理的版块列表（仅版主显示）
        items:
          $ref: '#/definitions/schema.CategoryBasicInfo'
        type: array
      points:
        description: Points | 积分
        example: 1000
        type: integer
      post_count:
        description: Post count | 帖子数
        example: 50
        type: integer
      readme:
        description: README content | README内容
        example: |-
          ## 关于我
          这是我的README内容
        type: string
      role:
        description: User role | 用户身份
        example: User
        type: string
      signature:
        description: Signature | 个性签名
        example: 这是我的个性签名
        type: string
      status:
        description: User status | 用户状态
        example: Normal
        type: string
      updated_at:
        description: Update time | 更新时间
        example: "2024-01-01 00:00:00"
        type: string
      username:
        description: Username | 用户名
        example: testuser
        type: string
    type: object
...
  schema.UserFollowItem:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      followed_at:
        example: "2024-01-20 12:00:00"
        type: string
      is_following:
        description: 当前用户是否关注此人
        example: true
        type: boolean
      is_mutual:
        description: 是否互相关注
        example: false
        type: boolean
      signature:
        example: 这是个性签名
        type: string
      user_id:
        example: 123
        type: integer
      username:
        example: username
        type: string
    type: object
  schema.UserFollowRequest:
    properties:
      following_id:
        example: 123
        minimum: 1
        type: integer
    required:
    - following_id
    type: object
  schema.UserFollowResponse:
    properties:
      following_id:
        example: 123
        type: integer
      message:
        example: 关注成功
        type: string
      success:
        example: true
        type: boolean
    type: object
  schema.UserFollowStatusResponse:
    properties:
      is_follower:
        description: 目标用户是否关注当前用户
        example: false
        type: boolean
      is_following:
        description: 当前用户是否关注目标用户
        example: true
        type: boolean
      is_mutual:
        description: 是否互相关注
        example: false
        type: boolean
    type: object
  schema.UserFollowersResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/schema.UserFollowItem'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 100
        type: integer
    type: object
  schema.UserFollowingResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/schema.UserFollowItem'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 50
        type: integer
    type: object
...
  schema.UserListItem:
    properties:
      avatar:
        description: Avatar URL | 头像URL
        example: https://example.com/avatar.jpg
        type: string
      comment_count:
        description: Comment count | 评论数
        example: 200
        type: integer
      created_at:
        description: Creation time | 创建时间
        example: "2024-01-01 00:00:00"
        type: string
      currency:
        description: Currency | 货币
        example: 500
        type: integer
      email:
        description: Email | 邮箱
        example: test@example.com
        type: string
      email_verified:
        description: Whether email verified | 邮箱是否已验证
        example: true
        type: boolean
      id:
        description: User ID | 用户ID
        example: 1
        type: integer
      points:
        description: Points | 积分
        example: 1000
        type: integer
      post_count:
        description: Post count | 帖子数
        example: 50
        type: integer
      role:
        description: User role | 用户身份
        example: User
        type: string
      signature:
        description: Signature | 个性签名
        example: 这是我的个性签名
        type: string
      status:
        description: User status | 用户状态
        example: Normal
        type: string
      updated_at:
        description: Update time | 更新时间
        example: "2024-01-01 00:00:00"
        type: string
      username:
        description: Username | 用户名
        example: testuser
        type: string
    type: object
  schema.UserListResponse:
    properties:
      list:
        description: User list | 用户列表
        items:
          $ref: '#/definitions/schema.UserListItem'
        type: array
      page:
        description: Current page number | 当前页码
        type: integer
      page_size:
        description: Items per page | 每页数量
        type: integer
      total:
        description: Total count | 总数量
        type: integer
    type: object
...
  schema.UserUnfollowResponse:
    properties:
      following_id:
        example: 123
        type: integer
      message:
        example: 取消关注成功
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  schema.UserUpdateRequest:
    properties:
      avatar:
        description: Avatar URL | 头像URL
        example: https://example.com/avatar.jpg
        type: string
      email:
        description: Email | 邮箱
        example: test@example.com
        type: string
      id:
        description: User ID | 用户ID
        example: 1
        type: integer
      readme:
        description: README content | README内容
        example: |-
          ## 关于我
          这是我的README内容
        type: string
      signature:
        description: Signature | 个性签名
        example: 这是我的个性签名
        type: string
      username:
        description: Username | 用户名
        example: testuser
        maxLength: 100
        minLength: 3
        type: string
    required:
    - id
    type: object
...
host: localhost:9876
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: PokeForum is a forum application based on Gin framework | PokeForum
    是一个基于 Gin 框架的论坛应用程序
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PokeForum API
  version: "1.0"
paths:
...
  /manage/users:
    get:
      consumes:
      - application/json
      description: Get paginated user list with keyword search and status filtering
        support | 分页获取用户列表，支持关键词搜索和状态筛选
      parameters:
      - description: Page number | 页码
        in: query
        name: page
        required: true
        type: integer
      - description: Items per page | 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      - description: Search keyword | 搜索关键词
        example: '"test"'
        in: query
        name: keyword
        type: string
      - description: User status | 用户状态
        example: '"Normal"'
        in: query
        name: status
        type: string
      - description: User role | 用户身份
        example: '"User"'
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieve successful | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserListResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get user list | 获取用户列表
      tags:
      - '[Admin]User Management | [管理员]用户管理'
    post:
      consumes:
      - application/json
      description: Administrator creates new user account | 管理员创建新用户账户
      parameters:
      - description: User information | 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Creation successful | 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserDetailResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Create user | 创建用户
      tags:
      - '[Admin]User Management | [管理员]用户管理'
    put:
      consumes:
      - application/json
      description: Update user's basic information | 更新用户的基本信息
      parameters:
      - description: User information | 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update successful | 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserDetailResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Update user information | 更新用户信息
      tags:
      - '[Admin]User Management | [管理员]用户管理'
  /manage/users/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information for specified user | 获取指定用户的详细信息
      parameters:
      - description: User ID | 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Retrieve successful | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserDetailResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get user details | 获取用户详情
      tags:
      - '[Admin]User Management | [管理员]用户管理'
...
  /profile/follow:
    post:
      consumes:
      - application/json
      description: Follow a specified user | 关注指定用户
      parameters:
      - description: Follow user request | 关注用户请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserFollowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Followed successfully | 关注成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserFollowResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Internal server error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Follow user | 关注用户
      tags:
      - '[User] User Follow | [用户个人中心] 用户关注'
  /profile/follow/{user_id}:
    delete:
      consumes:
      - application/json
      description: Unfollow a specified user | 取消关注指定用户
      parameters:
      - description: User ID to unfollow | 要取消关注的用户ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Unfollowed successfully | 取消关注成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserUnfollowResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Internal server error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Unfollow user | 取消关注用户
      tags:
      - '[User] User Follow | [用户个人中心] 用户关注'
  /profile/follow/followers:
    get:
      consumes:
      - application/json
      description: Get the specified user's followers list, supports pagination. If
        user_id is not provided, returns current user's followers | 获取指定用户的粉丝列表，支持分页。如果不提供user_id，则返回当前登录用户的粉丝列表
      parameters:
      - description: User ID, default to current user | 用户ID，默认为当前用户
        in: query
        name: user_id
        type: integer
      - default: 1
        description: Page number | 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page | 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserFollowersResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Internal server error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get user followers list | 获取用户粉丝列表
      tags:
      - '[User] User Follow | [用户个人中心] 用户关注'
  /profile/follow/following:
    get:
      consumes:
      - application/json
      description: Get the specified user's following list, supports pagination. If
        user_id is not provided, returns current user's following | 获取指定用户的关注列表，支持分页。如果不提供user_id，则返回当前登录用户的关注列表
      parameters:
      - description: User ID, default to current user | 用户ID，默认为当前用户
        in: query
        name: user_id
        type: integer
      - default: 1
        description: Page number | 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page | 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserFollowingResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Internal server error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get user following list | 获取用户关注列表
      tags:
      - '[User] User Follow | [用户个人中心] 用户关注'
  /profile/follow/status/{user_id}:
    get:
      consumes:
      - application/json
      description: Get the follow status between current user and specified user |
        获取当前用户与指定用户之间的关注状态
      parameters:
      - description: Target user ID | 目标用户ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserFollowStatusResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Internal server error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get follow status | 获取关注状态
      tags:
      - '[User] User Follow | [用户个人中心] 用户关注'
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"